move actions/view.py to peaksviewers/
