+++ qt/python/mantidqt/widgets/sliceviewer/peaksviewer/actions.py
