--- qt/python/mantidqt/widgets/sliceviewer/peaksviewer/actions/__init__.py
@@ -1,7 +1,0 @@
-# Mantid Repository : https://github.com/mantidproject/mantid
-#
-# Copyright &copy; 2020 ISIS Rutherford Appleton Laboratory UKRI,
-#     NScD Oak Ridge National Laboratory, European Spallation Source
-#     & Institut Laue - Langevin
-# SPDX - License - Identifier: GPL - 3.0 +
-#  This file is part of the mantid workbench.
